moves e2e tests for data entry api error responses to separate file
